# Fix AOT-related issues using function pointers