@adams85 2022-10-08
I made some further research and came to the conclusion that we'd better scrap this idea.

As mentioned in the feature documentation, function pointers are for static methods only, instance methods (or more precisely, the EXPLICITTHIS calling convention) are not supported directly. Although it would probably work in our case but would be considered a risky hack because we'd rely on unspecified behavior. ([This article](https://minidump.net/c-why-function-pointers-cant-be-used-on-instance-methods-8a99fc99b040) explains the details nicely.)

We could still use function pointers by converting the cached instance methods to static ones and adding an explicit this parameter but that would be too much pain for too little gain.